Use read/write fences for code load/execution
Rustc 1.21 supports read/write fences so this mechanism is now used as
an attempt to ensure that the PRU code is actually physically written to
memory before the code is executed with PruCode::run().

This was previously tentatively achieved by preventing the inlining of
PruLoader::load_code(), though I admit it is not clear to me whether
either solution is reliable or even necessary.