Port examples to the clpru assembler
The binary PRU code in the examples can now be built with either the
venerable pasm compiler or with the clpru compiler. Since the assembler
syntax is slightly different between these 2, the assembler sources are
in separate directories.

Since pasm is no longer maintained and clpru comes now installed by
default on the debian beaglebone distribution, it makes sense to move to
the latter.
This also makes prusst now completely independent from the apparently
unmaintained AM335x PRU package.